+++ barretenberg/cpp/src/barretenberg/common/thread.cpp
@@ -86,24 +86,3 @@ void parallel_for(size_t num_iterations, const std::function<void(size_t)>& func
 #endif
 #endif
 }
-
-/**
- * A modified parallel_for optimized for work being done in batches.
- * This is more appropriate for work with small granularity, to avoid thread caching issues and overhead.
- */
-inline void parallel_for_batched(size_t num_iterations, auto&& func)
-{
-    size_t num_threads = get_num_cpus_pow2();
-    size_t batch_size = (num_iterations + num_threads - 1) / num_threads; // round up division
-
-    // We will use parallel_for to dispatch the batches
-    parallel_for(num_threads, [=](size_t thread_idx) {
-        // Calculate start and end for this batch
-        size_t start = thread_idx * batch_size;
-        size_t end = std::min(start + batch_size, num_iterations);
-
-        for (size_t i = start; i < end; ++i) {
-            func(i); // Pass both the current index and thread_idx
-        }
-    });
-}
