+++ barretenberg/cpp/src/barretenberg/honk/pcs/zeromorph/zeromorph.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include "barretenberg/common/thread.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
 
 namespace proof_system::honk::pcs::zeromorph {
@@ -70,7 +69,7 @@ template <typename Curve> class ZeroMorphProver_ {
         }
 
         // Compute the q_k in reverse order, i.e. q_{n-1}, ..., q_0
-        parallel_for(log_N, [&](size_t k) {
+        for (size_t k = 0; k < log_N; ++k) {
             // Define partial evaluation point u' = (u_k, ..., u_{n-1})
             auto evaluation_point_size = static_cast<std::ptrdiff_t>(k + 1);
             std::vector<FF> u_partial(u_challenge.end() - evaluation_point_size, u_challenge.end());
@@ -78,18 +77,18 @@ template <typename Curve> class ZeroMorphProver_ {
             // Compute f' = f(X_0,...,X_{k-1}, u')
             auto f_1 = polynomial.partial_evaluate_mle(u_partial);
 
-            // // Increment first element to get altered partial evaluation point u'' = (u_k + 1, u_{k+1}, ..., u_{n-1})
-            // u_partial[0] += 1;
+            // Increment first element to get altered partial evaluation point u'' = (u_k + 1, u_{k+1}, ..., u_{n-1})
+            u_partial[0] += 1;
 
-            // // Compute f'' = f(X_0,...,X_{k-1}, u'')
-            // auto f_2 = polynomial.partial_evaluate_mle(u_partial);
+            // Compute f'' = f(X_0,...,X_{k-1}, u'')
+            auto f_2 = polynomial.partial_evaluate_mle(u_partial);
 
-            // // Compute q_k = f''(X_0,...,X_{k-1}) - f'(X_0,...,X_{k-1})
-            // auto q_k = f_2;
-            // q_k -= f_1;
+            // Compute q_k = f''(X_0,...,X_{k-1}) - f'(X_0,...,X_{k-1})
+            auto q_k = f_2;
+            q_k -= f_1;
 
-            // quotients[log_N - k - 1] = q_k;
-        });
+            quotients[log_N - k - 1] = q_k;
+        }
 
         return quotients;
     }
